new strategy for pfams
ok so i was looking for more ways but things that i made usually timed out on the google colab time limit so i made this thing that saves everything to google drive. however there still are some flaws to this as there are over 2k files to download and depending on the size you might not have enough storage for it... and also this is just as slow and probably you wont be able to get it all done in one run of the program, however, you can always just edit the excel spreadhsheet so the program picks up where you left off last execution based on the logs in the output console. i will be trying to get everything and then annotating everything using things like hmmer to get the pfams for more training